33_Docker_Compose_Demo/app/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
redis==4.6.0
requests==2.31.0
